refactor: moved `FooterTopShadow`
