+++ ts/screens/services/ServiceDetailsScreen.tsx
@@ -21,7 +21,7 @@ import ServiceMetadataComponent from "../../components/services/ServiceMetadata"
 import SpecialServicesCTA from "../../components/services/SpecialServices/SpecialServicesCTA";
 import TosAndPrivacyBox from "../../components/services/TosAndPrivacyBox";
 import Markdown from "../../components/ui/Markdown";
-import { FooterTopShadow } from "../../features/bonus/bonusVacanze/components/FooterTopShadow";
+import { FooterTopShadow } from "../../components/FooterTopShadow";
 import I18n from "../../i18n";
 import { IOStackNavigationRouteProps } from "../../navigation/params/AppParamsList";
 import { ServicesParamsList } from "../../navigation/params/ServicesParamsList";
